+++ stoix/configs/logger/base_logger.yaml
@@ -4,11 +4,11 @@ base_exp_path: results # Base path for logging.
 use_console: True # Whether to log to stdout.
 use_tb: False # Whether to use tensorboard logging.
 use_json: False # Whether to log marl-eval style to json files.
-use_neptune: False  # Whether to log to neptune.ai.
+use_neptune: True  # Whether to log to neptune.ai.
 
 # --- Other logger kwargs ---
 kwargs:
-  neptune_project: ~  # Project name in neptune.ai
+  neptune_project: InstaDeep/flashbax  # Project name in neptune.ai
   neptune_tag: [stoix]
   detailed_neptune_logging: False  # having mean/std/min/max can clutter neptune so we make it optional
   json_path: ~ # If set, json files will be logged to a set path so that multiple experiments can
