+++ stoix/configs/network/mlp_continuous.yaml
@@ -2,7 +2,7 @@
 actor_network:
   pre_torso:
     _target_: stoix.networks.torso.MLPTorso
-    layer_sizes: [64, 64]
+    layer_sizes: [256, 256]
     use_layer_norm: False
     activation: relu
   action_head:
@@ -11,7 +11,7 @@ actor_network:
 critic_network:
   pre_torso:
     _target_: stoix.networks.torso.MLPTorso
-    layer_sizes: [64, 64]
+    layer_sizes: [256, 256]
     use_layer_norm: False
     activation: relu
   critic_head:
