+++ stoix/configs/network/sampled_muzero.yaml
@@ -1,0 +1,43 @@
+# ---MLP Actor Critic Networks---
+actor_network:
+  pre_torso:
+    _target_: stoix.networks.torso.MLPTorso
+    layer_sizes: [256, 256]
+    use_layer_norm: False
+    activation: relu
+  action_head:
+    _target_: stoix.networks.heads.NormalAffineTanhDistributionHead
+
+critic_network:
+  pre_torso:
+    _target_: stoix.networks.torso.MLPTorso
+    layer_sizes: [256, 256]
+    use_layer_norm: False
+    activation: relu
+  critic_head:
+    _target_: stoix.networks.heads.CategoricalCriticHead
+
+# ---MLP MuZero Networks---
+representation_network:
+  torso:
+    _target_: stoix.networks.torso.MLPTorso
+    layer_sizes: [256, 256]
+    use_layer_norm: False
+    activation: relu
+  embedding_head:
+    _target_: stoix.networks.heads.LinearHead
+    output_dim: 128 # Output dimension of the embedding head. This should match the output dimension of the dynamics network.
+
+dynamics_network:
+  input_layer:
+    _target_: stoix.networks.inputs.EmbeddingActionInput
+  torso:
+    _target_: stoix.networks.torso.MLPTorso
+    layer_sizes: [256, 256]
+    use_layer_norm: False
+    activation: relu
+  embedding_head:
+    _target_: stoix.networks.heads.LinearHead
+    output_dim: 128 # Output dimension of the embedding head. This should match the output dimension of the representation network.
+  reward_head:
+    _target_: stoix.networks.heads.CategoricalCriticHead
