--- stoix/systems/search/ff_sampled_az.py
@@ -1,718 +1,0 @@
-import copy
-import functools
-import time
-from typing import Any, Dict, Tuple
-
-import chex
-import flax
-import hydra
-import jax
-import jax.numpy as jnp
-import mctx
-import optax
-import rlax
-import tensorflow_probability.substrates.jax as tfp
-from colorama import Fore, Style
-from flax.core.frozen_dict import FrozenDict
-from jumanji.env import Environment
-from omegaconf import DictConfig, OmegaConf
-from rich.pretty import pprint
-
-from stoix.base_types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
-from stoix.networks.base import FeedForwardActor as Actor
-from stoix.networks.base import FeedForwardCritic as Critic
-from stoix.systems.ppo.ppo_types import (
-    ActorCriticOptStates,
-    ActorCriticParams,
-    LearnerState,
-)
-from stoix.systems.search.evaluator import search_evaluator_setup
-from stoix.systems.search.search_types import (
-    EnvironmentStep,
-    RootFnApply,
-    SampledExItTransition,
-    SearchApply,
-)
-from stoix.utils import make_env as environments
-from stoix.utils.checkpointing import Checkpointer
-from stoix.utils.jax_utils import (
-    merge_leading_dims,
-    unreplicate_batch_dim,
-    unreplicate_n_dims,
-)
-from stoix.utils.logger import LogEvent, StoixLogger
-from stoix.utils.loss import clipped_value_loss
-from stoix.utils.multistep import batch_truncated_generalized_advantage_estimation
-from stoix.utils.total_timestep_checker import check_total_timesteps
-from stoix.utils.training import make_learning_rate
-from stoix.wrappers.episode_metrics import get_final_step_metrics
-
-tfd = tfp.distributions
-
-
-def make_sampled_search_apply_fn(
-    config: DictConfig,
-    model_recurrent_fn: mctx.RecurrentFn,
-) -> SearchApply:
-
-    search_method = parse_search_method(config)
-    mctx_search_apply_fn = functools.partial(
-        search_method,
-        recurrent_fn=model_recurrent_fn,
-        num_simulations=config.system.num_simulations,
-        max_depth=config.system.max_depth,
-    )
-
-    def search_apply_fn(
-        params: ActorCriticParams,
-        key: chex.PRNGKey,
-        root: mctx.RootFnOutput,
-    ) -> mctx.PolicyOutput:
-        """Apply the search function."""
-        root_sampled_actions = root.embedding["sampled_actions"]
-        search_output = mctx_search_apply_fn(params, key, root)
-        action_index = search_output.action
-        batch_indices = jnp.arange(action_index.shape[0])
-        action = root_sampled_actions[batch_indices, action_index]
-        chex.assert_shape(action, (action_index.shape[0], config.system.action_dim))
-        return search_output.replace(action=action)
-
-    return search_apply_fn
-
-
-def make_root_fn(
-    actor_apply_fn: ActorApply, critic_apply_fn: CriticApply, config: DictConfig
-) -> RootFnApply:
-    def root_fn(
-        params: ActorCriticParams,
-        observation: chex.ArrayTree,
-        state_embedding: chex.ArrayTree,
-        rng_key: chex.PRNGKey,
-    ) -> mctx.RootFnOutput:
-
-        sample_key, noise_key = jax.random.split(rng_key, 2)
-
-        # Calculate the value and policy for the root node.
-        pi = actor_apply_fn(params.actor_params, observation)
-        value = critic_apply_fn(params.critic_params, observation)
-
-        batch_size = value.shape[0]
-        # Sample actions for the root node.
-        sampled_actions = pi.sample(seed=sample_key, sample_shape=config.system.num_samples)
-        chex.assert_shape(
-            sampled_actions, (config.system.num_samples, batch_size, config.system.action_dim)
-        )
-        # Swap axes to have (batch_size, num_samples, action_dim)
-        sampled_actions = jnp.swapaxes(sampled_actions, 0, 1)
-        chex.assert_shape(
-            sampled_actions, (batch_size, config.system.num_samples, config.system.action_dim)
-        )
-        # Add noise to the root sampled actions.
-        if config.system.root_exploration_sigma != 0:
-            sampled_actions = rlax.add_gaussian_noise(
-                noise_key, sampled_actions, config.system.root_exploration_sigma
-            )
-        # Due to sampling, set all actions to have a uniform prior.
-        selection_logits = jnp.ones((batch_size, config.system.num_samples))
-        # Pack the search tree state.
-        search_tree_state = {"env_state": state_embedding, "sampled_actions": sampled_actions}
-
-        root_fn_output = mctx.RootFnOutput(
-            prior_logits=selection_logits,
-            value=value,
-            embedding=search_tree_state,
-        )
-
-        return root_fn_output
-
-    return root_fn
-
-
-def make_recurrent_fn(
-    environment_step: EnvironmentStep,
-    actor_apply_fn: ActorApply,
-    critic_apply_fn: CriticApply,
-    config: DictConfig,
-) -> mctx.RecurrentFn:
-    def recurrent_fn(
-        params: ActorCriticParams,
-        rng_key: chex.PRNGKey,
-        action_index: chex.Array,
-        search_tree_state: chex.ArrayTree,
-    ) -> Tuple[mctx.RecurrentFnOutput, chex.ArrayTree]:
-
-        batch_size = action_index.shape[0]
-        batch_indices = jnp.arange(batch_size)
-        # Get the sampled actions from the search tree state.
-        sampled_actions = search_tree_state["sampled_actions"]
-        # Select the action from the sampled actions.
-        action = sampled_actions[batch_indices, action_index]
-        # Get the environment state from the search tree state.
-        state_embedding = search_tree_state["env_state"]
-        # Step the environment.
-        next_state_embedding, next_timestep = environment_step(state_embedding, action)
-
-        # Calculate the value and policy for the next state.
-        pi = actor_apply_fn(params.actor_params, next_timestep.observation)
-        value = critic_apply_fn(params.critic_params, next_timestep.observation)
-
-        # Sample actions for the next state.
-        next_sampled_actions = pi.sample(seed=rng_key, sample_shape=config.system.num_samples)
-        chex.assert_shape(
-            next_sampled_actions, (config.system.num_samples, batch_size, config.system.action_dim)
-        )
-        # Swap axes to have (batch_size, num_samples, action_dim)
-        next_sampled_actions = jnp.swapaxes(next_sampled_actions, 0, 1)
-        chex.assert_shape(
-            next_sampled_actions, (batch_size, config.system.num_samples, config.system.action_dim)
-        )
-        # Due to sampling, set all actions to have a uniform prior.
-        selection_logits = jnp.ones((batch_size, config.system.num_samples))
-        # Pack the search tree state.
-        next_search_tree_state = {
-            "env_state": next_state_embedding,
-            "sampled_actions": next_sampled_actions,
-        }
-
-        recurrent_fn_output = mctx.RecurrentFnOutput(
-            reward=next_timestep.reward,
-            discount=next_timestep.discount * config.system.gamma,
-            prior_logits=selection_logits,
-            value=next_timestep.discount * value,
-        )
-
-        return recurrent_fn_output, next_search_tree_state
-
-    return recurrent_fn
-
-
-def get_learner_fn(
-    env: Environment,
-    apply_fns: Tuple[ActorApply, CriticApply, RootFnApply, SearchApply],
-    update_fns: Tuple[optax.TransformUpdateFn, optax.TransformUpdateFn],
-    config: DictConfig,
-) -> LearnerFn[LearnerState]:
-    """Get the learner function."""
-
-    # Get apply and update functions for actor and critic networks.
-    actor_apply_fn, critic_apply_fn, root_fn, search_apply_fn = apply_fns
-    actor_update_fn, critic_update_fn = update_fns
-
-    def _update_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, Tuple]:
-        """A single update of the network."""
-
-        def _env_step(
-            learner_state: LearnerState, _: Any
-        ) -> Tuple[LearnerState, SampledExItTransition]:
-            """Step the environment."""
-            params, opt_states, key, env_state, last_timestep = learner_state
-
-            # SELECT ACTION
-            key, root_key, policy_key = jax.random.split(key, num=3)
-            root = root_fn(params, last_timestep.observation, env_state.env_state, root_key)
-            root_sampled_actions = root.embedding["sampled_actions"]
-            search_output = search_apply_fn(params, policy_key, root)
-            action = search_output.action
-            search_policy = search_output.action_weights
-            search_value = search_output.search_tree.node_values[:, mctx.Tree.ROOT_INDEX]
-            behaviour_value = critic_apply_fn(params.critic_params, last_timestep.observation)
-
-            # STEP ENVIRONMENT
-            env_state, timestep = jax.vmap(env.step, in_axes=(0, 0))(env_state, action)
-
-            # LOG EPISODE METRICS
-            done = timestep.last().reshape(-1)
-            info = timestep.extras["episode_metrics"]
-
-            transition = SampledExItTransition(
-                done,
-                action,
-                behaviour_value,
-                timestep.reward,
-                search_value,
-                search_policy,
-                root_sampled_actions,
-                last_timestep.observation,
-                info,
-            )
-            learner_state = LearnerState(params, opt_states, key, env_state, timestep)
-            return learner_state, transition
-
-        # STEP ENVIRONMENT FOR ROLLOUT LENGTH
-        learner_state, traj_batch = jax.lax.scan(
-            _env_step, learner_state, None, config.system.rollout_length
-        )
-
-        # CALCULATE ADVANTAGE
-        params, opt_states, key, env_state, last_timestep = learner_state
-        last_val = critic_apply_fn(params.critic_params, last_timestep.observation)
-
-        r_t = traj_batch.reward
-        v_t = jnp.concatenate([traj_batch.value, last_val[None, ...]], axis=0)
-        d_t = 1.0 - traj_batch.done.astype(jnp.float32)
-        d_t = (d_t * config.system.gamma).astype(jnp.float32)
-        _, targets = batch_truncated_generalized_advantage_estimation(
-            r_t, d_t, config.system.gae_lambda, v_t, time_major=True
-        )
-
-        def _update_epoch(update_state: Tuple, _: Any) -> Tuple:
-            """Update the network for a single epoch."""
-
-            def _update_minibatch(train_state: Tuple, batch_info: Tuple) -> Tuple:
-                """Update the network for a single minibatch."""
-
-                # UNPACK TRAIN STATE AND BATCH INFO
-                params, opt_states, key = train_state
-                traj_batch, targets = batch_info
-
-                def _actor_loss_fn(
-                    actor_params: FrozenDict,
-                    traj_batch: SampledExItTransition,
-                    rng_key: chex.PRNGKey,
-                ) -> Tuple:
-                    """Calculate the actor loss."""
-                    # RERUN NETWORK
-                    actor_policy = actor_apply_fn(actor_params, traj_batch.obs)
-
-                    # CALCULATE LOSS
-                    actor_log_probs = jax.vmap(actor_policy.log_prob, in_axes=1, out_axes=1)(
-                        traj_batch.sampled_actions
-                    )
-                    actor_loss = -jnp.sum(
-                        actor_log_probs * traj_batch.search_policy, axis=-1
-                    ).mean()
-
-                    entropy = actor_policy.entropy(seed=rng_key).mean()
-
-                    total_loss_actor = actor_loss - config.system.ent_coef * entropy
-                    return total_loss_actor, (actor_loss, entropy)
-
-                def _critic_loss_fn(
-                    critic_params: FrozenDict,
-                    traj_batch: SampledExItTransition,
-                    targets: chex.Array,
-                ) -> Tuple:
-                    """Calculate the critic loss."""
-                    # RERUN NETWORK
-                    value = critic_apply_fn(critic_params, traj_batch.obs)
-
-                    # CALCULATE VALUE LOSS
-                    value_loss = clipped_value_loss(
-                        value, traj_batch.value, targets, config.system.clip_eps
-                    )
-
-                    critic_total_loss = config.system.vf_coef * value_loss
-                    return critic_total_loss, (value_loss)
-
-                # CALCULATE ACTOR LOSS
-                key, actor_loss_key = jax.random.split(key)
-                actor_grad_fn = jax.value_and_grad(_actor_loss_fn, has_aux=True)
-                actor_loss_info, actor_grads = actor_grad_fn(
-                    params.actor_params, traj_batch, actor_loss_key
-                )
-
-                # CALCULATE CRITIC LOSS
-                critic_grad_fn = jax.value_and_grad(_critic_loss_fn, has_aux=True)
-                critic_loss_info, critic_grads = critic_grad_fn(
-                    params.critic_params, traj_batch, targets
-                )
-
-                # Compute the parallel mean (pmean) over the batch.
-                # This calculation is inspired by the Anakin architecture demo notebook.
-                # available at https://tinyurl.com/26tdzs5x
-                # This pmean could be a regular mean as the batch axis is on the same device.
-                actor_grads, actor_loss_info = jax.lax.pmean(
-                    (actor_grads, actor_loss_info), axis_name="batch"
-                )
-                # pmean over devices.
-                actor_grads, actor_loss_info = jax.lax.pmean(
-                    (actor_grads, actor_loss_info), axis_name="device"
-                )
-
-                critic_grads, critic_loss_info = jax.lax.pmean(
-                    (critic_grads, critic_loss_info), axis_name="batch"
-                )
-                # pmean over devices.
-                critic_grads, critic_loss_info = jax.lax.pmean(
-                    (critic_grads, critic_loss_info), axis_name="device"
-                )
-
-                # UPDATE ACTOR PARAMS AND OPTIMISER STATE
-                actor_updates, actor_new_opt_state = actor_update_fn(
-                    actor_grads, opt_states.actor_opt_state
-                )
-                actor_new_params = optax.apply_updates(params.actor_params, actor_updates)
-
-                # UPDATE CRITIC PARAMS AND OPTIMISER STATE
-                critic_updates, critic_new_opt_state = critic_update_fn(
-                    critic_grads, opt_states.critic_opt_state
-                )
-                critic_new_params = optax.apply_updates(params.critic_params, critic_updates)
-
-                # PACK NEW PARAMS AND OPTIMISER STATE
-                new_params = ActorCriticParams(actor_new_params, critic_new_params)
-                new_opt_state = ActorCriticOptStates(actor_new_opt_state, critic_new_opt_state)
-
-                # PACK LOSS INFO
-                total_loss = actor_loss_info[0] + critic_loss_info[0]
-                value_loss = critic_loss_info[1]
-                actor_loss = actor_loss_info[1][0]
-                entropy = actor_loss_info[1][1]
-                loss_info = {
-                    "total_loss": total_loss,
-                    "value_loss": value_loss,
-                    "actor_loss": actor_loss,
-                    "entropy": entropy,
-                }
-                return (new_params, new_opt_state, key), loss_info
-
-            params, opt_states, traj_batch, targets, key = update_state
-            key, shuffle_key = jax.random.split(key)
-
-            # SHUFFLE MINIBATCHES
-            batch_size = config.system.rollout_length * config.arch.num_envs
-            permutation = jax.random.permutation(shuffle_key, batch_size)
-            batch = (traj_batch, targets)
-            batch = jax.tree_util.tree_map(lambda x: merge_leading_dims(x, 2), batch)
-            shuffled_batch = jax.tree_util.tree_map(
-                lambda x: jnp.take(x, permutation, axis=0), batch
-            )
-            minibatches = jax.tree_util.tree_map(
-                lambda x: jnp.reshape(x, [config.system.num_minibatches, -1] + list(x.shape[1:])),
-                shuffled_batch,
-            )
-
-            # UPDATE MINIBATCHES
-            (params, opt_states, key), loss_info = jax.lax.scan(
-                _update_minibatch, (params, opt_states, key), minibatches
-            )
-
-            update_state = (params, opt_states, traj_batch, targets, key)
-            return update_state, loss_info
-
-        update_state = (params, opt_states, traj_batch, targets, key)
-
-        # UPDATE EPOCHS
-        update_state, loss_info = jax.lax.scan(
-            _update_epoch, update_state, None, config.system.epochs
-        )
-
-        params, opt_states, traj_batch, targets, key = update_state
-        learner_state = LearnerState(params, opt_states, key, env_state, last_timestep)
-        metric = traj_batch.info
-        return learner_state, (metric, loss_info)
-
-    def learner_fn(learner_state: LearnerState) -> ExperimentOutput[LearnerState]:
-        """Learner function."""
-
-        batched_update_step = jax.vmap(_update_step, in_axes=(0, None), axis_name="batch")
-
-        learner_state, (episode_info, loss_info) = jax.lax.scan(
-            batched_update_step, learner_state, None, config.arch.num_updates_per_eval
-        )
-        return ExperimentOutput(
-            learner_state=learner_state,
-            episode_metrics=episode_info,
-            train_metrics=loss_info,
-        )
-
-    return learner_fn
-
-
-def parse_search_method(config: DictConfig) -> Any:
-    """Parse search method from config."""
-    if config.system.search_method.lower() == "muzero":
-        # Set the search method to MuZero with no dirichlet noise.
-        search_method = functools.partial(
-            mctx.muzero_policy, dirichlet_fraction=0.0, dirichlet_alpha=0.0
-        )
-    elif config.system.search_method.lower() == "gumbel":
-        search_method = mctx.gumbel_muzero_policy
-    else:
-        raise ValueError(f"Search method {config.system.search_method} not supported.")
-
-    return search_method
-
-
-def learner_setup(
-    env: Environment, keys: chex.Array, config: DictConfig, model_env: Environment
-) -> Tuple[LearnerFn[LearnerState], RootFnApply, SearchApply, LearnerState]:
-    """Initialise learner_fn, network, optimiser, environment and states."""
-    # Get available TPU cores.
-    n_devices = len(jax.devices())
-
-    # Get number of actions.
-    num_actions = int(env.action_spec().shape[-1])
-    config.system.action_dim = num_actions
-    config.system.action_minimum = float(env.action_spec().minimum)
-    config.system.action_maximum = float(env.action_spec().maximum)
-
-    # PRNG keys.
-    key, actor_net_key, critic_net_key = keys
-
-    # Define network and optimiser.
-    actor_torso = hydra.utils.instantiate(config.network.actor_network.pre_torso)
-    actor_action_head = hydra.utils.instantiate(
-        config.network.actor_network.action_head,
-        action_dim=num_actions,
-        minimum=env.action_spec().minimum,
-        maximum=env.action_spec().maximum,
-    )
-    critic_torso = hydra.utils.instantiate(config.network.critic_network.pre_torso)
-    critic_head = hydra.utils.instantiate(config.network.critic_network.critic_head)
-
-    actor_network = Actor(torso=actor_torso, action_head=actor_action_head)
-    critic_network = Critic(torso=critic_torso, critic_head=critic_head)
-
-    actor_lr = make_learning_rate(
-        config.system.actor_lr, config, config.system.epochs, config.system.num_minibatches
-    )
-    critic_lr = make_learning_rate(
-        config.system.critic_lr, config, config.system.epochs, config.system.num_minibatches
-    )
-
-    actor_optim = optax.chain(
-        optax.clip_by_global_norm(config.system.max_grad_norm),
-        optax.adam(actor_lr, eps=1e-5),
-    )
-    critic_optim = optax.chain(
-        optax.clip_by_global_norm(config.system.max_grad_norm),
-        optax.adam(critic_lr, eps=1e-5),
-    )
-
-    # Initialise observation: Select only obs for a single agent.
-    init_x = env.observation_spec().generate_value()
-    init_x = jax.tree_util.tree_map(lambda x: x[None, ...], init_x)
-
-    # Initialise actor params and optimiser state.
-    actor_params = actor_network.init(actor_net_key, init_x)
-    actor_opt_state = actor_optim.init(actor_params)
-
-    # Initialise critic params and optimiser state.
-    critic_params = critic_network.init(critic_net_key, init_x)
-    critic_opt_state = critic_optim.init(critic_params)
-
-    # Pack params.
-    params = ActorCriticParams(actor_params, critic_params)
-
-    actor_network_apply_fn = actor_network.apply
-    critic_network_apply_fn = critic_network.apply
-
-    root_fn = make_root_fn(actor_network_apply_fn, critic_network_apply_fn, config)
-    environment_model_step = jax.vmap(model_env.step)
-    model_recurrent_fn = make_recurrent_fn(
-        environment_model_step, actor_network_apply_fn, critic_network_apply_fn, config
-    )
-    search_apply_fn = make_sampled_search_apply_fn(config, model_recurrent_fn)
-
-    # Pack apply and update functions.
-    apply_fns = (
-        actor_network_apply_fn,
-        critic_network_apply_fn,
-        root_fn,
-        search_apply_fn,
-    )
-    update_fns = (actor_optim.update, critic_optim.update)
-
-    # Get batched iterated update and replicate it to pmap it over cores.
-    learn = get_learner_fn(env, apply_fns, update_fns, config)
-    learn = jax.pmap(learn, axis_name="device")
-
-    # Initialise environment states and timesteps: across devices and batches.
-    key, *env_keys = jax.random.split(
-        key, n_devices * config.system.update_batch_size * config.arch.num_envs + 1
-    )
-    env_states, timesteps = jax.vmap(env.reset, in_axes=(0))(
-        jnp.stack(env_keys),
-    )
-    reshape_states = lambda x: x.reshape(
-        (n_devices, config.system.update_batch_size, config.arch.num_envs) + x.shape[1:]
-    )
-    # (devices, update batch size, num_envs, ...)
-    env_states = jax.tree_map(reshape_states, env_states)
-    timesteps = jax.tree_map(reshape_states, timesteps)
-
-    # Load model from checkpoint if specified.
-    if config.logger.checkpointing.load_model:
-        loaded_checkpoint = Checkpointer(
-            model_name=config.system.system_name,
-            **config.logger.checkpointing.load_args,  # Other checkpoint args
-        )
-        # Restore the learner state from the checkpoint
-        restored_params, _ = loaded_checkpoint.restore_params()
-        # Update the params
-        params = restored_params
-
-    # Define params to be replicated across devices and batches.
-    key, step_keys = jax.random.split(key)
-    opt_states = ActorCriticOptStates(actor_opt_state, critic_opt_state)
-    replicate_learner = (params, opt_states, step_keys)
-
-    # Duplicate learner for update_batch_size.
-    broadcast = lambda x: jnp.broadcast_to(x, (config.system.update_batch_size,) + x.shape)
-    replicate_learner = jax.tree_map(broadcast, replicate_learner)
-
-    # Duplicate learner across devices.
-    replicate_learner = flax.jax_utils.replicate(replicate_learner, devices=jax.devices())
-
-    # Initialise learner state.
-    params, opt_states, step_keys = replicate_learner
-    init_learner_state = LearnerState(params, opt_states, step_keys, env_states, timesteps)
-
-    return learn, root_fn, search_apply_fn, init_learner_state
-
-
-def run_experiment(_config: DictConfig) -> None:
-    """Runs experiment."""
-    config = copy.deepcopy(_config)
-
-    # Calculate total timesteps.
-    n_devices = len(jax.devices())
-    config = check_total_timesteps(config)
-    assert (
-        config.arch.num_updates > config.arch.num_evaluation
-    ), "Number of updates per evaluation must be less than total number of updates."
-
-    # Create the environments for train and eval.
-    env, eval_env = environments.make(config=config)
-
-    # PRNG keys.
-    key, key_e, actor_net_key, critic_net_key = jax.random.split(
-        jax.random.PRNGKey(config.arch.seed), num=4
-    )
-
-    # Setup learner.
-    learn, root_fn, search_apply_fn, learner_state = learner_setup(
-        env, (key, actor_net_key, critic_net_key), config, eval_env
-    )
-
-    # Setup evaluator.
-    evaluator, absolute_metric_evaluator, (trained_params, eval_keys) = search_evaluator_setup(
-        eval_env=eval_env,
-        key_e=key_e,
-        search_apply_fn=search_apply_fn,
-        root_fn=root_fn,
-        params=learner_state.params,
-        config=config,
-    )
-
-    # Calculate number of updates per evaluation.
-    config.arch.num_updates_per_eval = config.arch.num_updates // config.arch.num_evaluation
-    steps_per_rollout = (
-        n_devices
-        * config.arch.num_updates_per_eval
-        * config.system.rollout_length
-        * config.system.update_batch_size
-        * config.arch.num_envs
-    )
-
-    # Logger setup
-    logger = StoixLogger(config)
-    cfg: Dict = OmegaConf.to_container(config, resolve=True)
-    cfg["arch"]["devices"] = jax.devices()
-    pprint(cfg)
-
-    # Set up checkpointer
-    save_checkpoint = config.logger.checkpointing.save_model
-    if save_checkpoint:
-        checkpointer = Checkpointer(
-            metadata=config,  # Save all config as metadata in the checkpoint
-            model_name=config.system.system_name,
-            **config.logger.checkpointing.save_args,  # Checkpoint args
-        )
-
-    # Run experiment for a total number of evaluations.
-    max_episode_return = jnp.float32(-1e7)
-    best_params = unreplicate_batch_dim(learner_state.params)
-    for eval_step in range(config.arch.num_evaluation):
-        # Train.
-        start_time = time.time()
-
-        learner_output = learn(learner_state)
-        jax.block_until_ready(learner_output)
-
-        # Log the results of the training.
-        elapsed_time = time.time() - start_time
-        t = int(steps_per_rollout * (eval_step + 1))
-        episode_metrics, ep_completed = get_final_step_metrics(learner_output.episode_metrics)
-        episode_metrics["steps_per_second"] = steps_per_rollout / elapsed_time
-
-        # Separately log timesteps, actoring metrics and training metrics.
-        logger.log({"timestep": t}, t, eval_step, LogEvent.MISC)
-        if ep_completed:  # only log episode metrics if an episode was completed in the rollout.
-            logger.log(episode_metrics, t, eval_step, LogEvent.ACT)
-        logger.log(learner_output.train_metrics, t, eval_step, LogEvent.TRAIN)
-
-        # Prepare for evaluation.
-        start_time = time.time()
-        trained_params = unreplicate_batch_dim(
-            learner_output.learner_state.params
-        )  # Select only actor params
-        key_e, *eval_keys = jax.random.split(key_e, n_devices + 1)
-        eval_keys = jnp.stack(eval_keys)
-        eval_keys = eval_keys.reshape(n_devices, -1)
-
-        # Evaluate.
-        evaluator_output = evaluator(trained_params, eval_keys)
-        jax.block_until_ready(evaluator_output)
-
-        # Log the results of the evaluation.
-        elapsed_time = time.time() - start_time
-        episode_return = jnp.mean(evaluator_output.episode_metrics["episode_return"])
-
-        steps_per_eval = int(jnp.sum(evaluator_output.episode_metrics["episode_length"]))
-        evaluator_output.episode_metrics["steps_per_second"] = steps_per_eval / elapsed_time
-        logger.log(evaluator_output.episode_metrics, t, eval_step, LogEvent.EVAL)
-
-        if save_checkpoint:
-            # Save checkpoint of learner state
-            checkpointer.save(
-                timestep=int(steps_per_rollout * (eval_step + 1)),
-                unreplicated_learner_state=unreplicate_n_dims(learner_output.learner_state),
-                episode_return=episode_return,
-            )
-
-        if config.arch.absolute_metric and max_episode_return <= episode_return:
-            best_params = copy.deepcopy(trained_params)
-            max_episode_return = episode_return
-
-        # Update runner state to continue training.
-        learner_state = learner_output.learner_state
-
-    # Measure absolute metric.
-    if config.arch.absolute_metric:
-        start_time = time.time()
-
-        key_e, *eval_keys = jax.random.split(key_e, n_devices + 1)
-        eval_keys = jnp.stack(eval_keys)
-        eval_keys = eval_keys.reshape(n_devices, -1)
-
-        evaluator_output = absolute_metric_evaluator(best_params, eval_keys)
-        jax.block_until_ready(evaluator_output)
-
-        elapsed_time = time.time() - start_time
-        t = int(steps_per_rollout * (eval_step + 1))
-        steps_per_eval = int(jnp.sum(evaluator_output.episode_metrics["episode_length"]))
-        evaluator_output.episode_metrics["steps_per_second"] = steps_per_eval / elapsed_time
-        logger.log(evaluator_output.episode_metrics, t, eval_step, LogEvent.ABSOLUTE)
-
-    # Stop the logger.
-    logger.stop()
-
-
-@hydra.main(
-    config_path="../../configs", config_name="default_ff_sampled_az.yaml", version_base="1.2"
-)
-def hydra_entry_point(cfg: DictConfig) -> None:
-    """Experiment entry point."""
-    # Allow dynamic attributes.
-    OmegaConf.set_struct(cfg, False)
-
-    # Run experiment.
-    run_experiment(cfg)
-
-    print(f"{Fore.CYAN}{Style.BRIGHT}Sampled AZ experiment completed{Style.RESET_ALL}")
-
-
-if __name__ == "__main__":
-    hydra_entry_point()
