+++ stoix/systems/vpg/ff_reinforce_continuous.py
@@ -218,7 +218,7 @@ def learner_setup(
     # Get available TPU cores.
     n_devices = len(jax.devices())
 
-    # Get number of actions and agents.
+    # Get number/dimension of actions.
     num_actions = int(env.action_spec().shape[-1])
     config.system.action_dim = num_actions
     config.system.action_minimum = float(env.action_spec().minimum)
@@ -253,7 +253,7 @@ def learner_setup(
         optax.adam(critic_lr, eps=1e-5),
     )
 
-    # Initialise observation: Select only obs for a single agent.
+    # Initialise observation
     init_x = env.observation_spec().generate_value()
     init_x = jax.tree_util.tree_map(lambda x: x[None, ...], init_x)
